requirements.txt
aiohttp>=3.9.0
coinbase-advanced-py>=1.0.0
flask>=3.0.0
plaid-python>=22.0.0
python-dotenv>=1.0.0
requests>=2.28.0
pyyaml>=6.0.0
sqlite-utils>=3.36
click>=8.1
rich>=13.7
plotext>=5.2
python-dateutil>=2.8
google-api-python-client>=2.100
google-auth>=2.25
